requirements.txt
streamlit
streamlit-option-menu
pandas
sarvamai
groq
fpdf2
tensorflow
numpy
Pillow

